Extract install_flow tests into module file
Verify the finding against current code and apply only the needed\nrefactor: move the oversized inline install_flow test module into\na dedicated submodule file.\n\n- Replace inline #[cfg(test)] mod tests { ... } with #[cfg(test)]\n  mod tests;\n- Add installer/src/install_flow/tests.rs containing all moved\n  fixtures, stubs, rstest cases, and fallback regression test\n- Preserve existing test behaviour and private-item access via\n  use super::*
